backport:
SVN commit 916644 by osterfeld: 

fix Complete Story link if the link contains an anchor. We must tell 
KHTMLPart that it shows the linked item to make relative links work, 
but that in turn breaks in the presence of an anchor (as KHTMLPart 
tries to simply jump to the anchor instead of opening the website).

Workaround: adding a bogus 
query item makes the two URLs different for the anchor check but 
keeps the relative link resolution working. 

CCBUG:177754
CCBUG:180789 

svn path=/branches/KDE/4.2/kdepim/; revision=916646