merge from 4.3 branch ------------------------------------------------------------------------ r996539 | osterfeld | 2009-07-14 16:08:16 +0200 (Tue, 14 Jul 2009) | 11 lines
This patch enables ctrl-+ and ctrl-- again to enlarge/shrink the font of the
articleviewer (font != zoom). These two actions are also added to the "View"
menu.
Additionally actions for print (ctrl-p) and copy (ctrl-c) now also work again
with shortcuts and are configurable in the shortcut configuration menu.

Patch by Alexander Elbs

BUG: 163391


------------------------------------------------------------------------

svn path=/trunk/KDE/kdepim/akregator/; revision=1059092