Export articles (signals), plugins with parents and KCModules as plugins
This patch export the selected articles with a signal, enabling plugins to
know which articles are selected and perform actions.

It also set's akregator_part as the parent of the plugins, enabling them to
connect to it's signals and have a way of communicating with it (like getting
the signal of selected articles).

Also load the config modules looking for them on service trade, instead of
hardcoding: this way it's easier to provide config modules for plugins: it just
needs a proper .desktop file.

Reviewed-by: Frank

svn path=/trunk/KDE/kdepim/akregator/; revision=1174532