Remember Size of Title Column in ArticleListView
The size of the title column in the article list view of akregator
is reset with each restart or even change of feed to some middle
position. This is very annoying when reading many feeds on a smaller
screen, since the title's are mostly cut. This patch fixes this
behavior, only sets the size automatically when no is given and always
uses and remembers the size set by the user.

Patch by Torsten Grote <grote@kolabsys.com>