Deal with akregator part main widget being externally deleted.
Running akregator under valgrind (to track down a different crash), I
got that the main widget was deleted from the event loop (by
qDeleteInEventHandler), which then triggered the deletion of the
akregator part, which then tried to access the (now deleted) widget.

This should prevent the code from crashing in this case.

REVIEW: 125710