Several performance optimisations for Akregator
Summary:
1) Fetch title, pubdate, hash and status in a single storage call

Even with MK4, the performance gain is sensible.
For medium sized archives (~25k articles), it's feeling faster.
For huge archives (>100k articles), it is human measurably faster.

2) Reuse the FeedStorage instead of fetching it for each article

For huge archives, this helps performance a bit

3) Drop useless call to isNull in model

All calls are safe even when the article is null.
And their results will be similar to QVariant().
This call is thus useless, and one of the biggest 'perf' offender

Reviewers: mlaurent, knauss

Reviewed By: mlaurent

Subscribers: dvratil, #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D12480