Make the feed uri the baseUrl for previews
Summary:
For some feeds (linuxfr.org for instance), the links in the feed are
relative to the feed URI (for instance, //linuxfr.org/some/content).
These links were broken by the forced baseUri of file://. It is better to
instead use the feed uri as the base.

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D19739