Remove use of deprecated QModelIndex::child method
Summary:
This will reduce warnings when compiling, and likely make the transition to Qt 6 a tiny bit
simpler. The replacement code is as recommended by Qt documentation, and should be equivalent.

Test Plan:
Checked the unit tests pass and that basic operations (reorder, create a folder, move
entries between folders) still work in the subscription tree in the UI.

Reviewers: vkrause

Reviewed By: vkrause

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D23959