akregator: port away from ReadOnlyPart.
Summary:
This removes the konqueror integration which didn't really work anyway
(RMB on .rss file / Preview In / Akregator -> akregrator would be
embedded but wouldn't add this feed). RMB / Add to Akregator is simpler
and works better anyway.

This will allow us to load kontact plugins without desktop files
(using json) in the future.

Depends on D28581.

Test Plan:
akregator still works (standalone and in kontact) and I was able to add a new feed
in the GUI (and the icon worked btw, also after a restart).

Reviewers: mlaurent

Differential Revision: https://phabricator.kde.org/D28584