Re-register the MainWindow with the system tray icon if its window handle changes
This is initially set as the KStatusNotifierItem's associated window,
but further on during startup the window gets destroyed and recreated.
This happens during the initialisation of the article viewer, somewhere
deep within QWebEngine. Since the KStatusNotifierItem retains the
associated window using its window handle, that association is lost.
The symptom is that the tray icon menu does not have a "Restore" item
and clicking on the icon does not open the main window.

The event filter monitors the MainWindow's window handle and, if it changes
(either during startup or while running) for any reason, re-registers it as
the tray icon's associated window.

BUG:494071